% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaBenchmark.R
\name{adaBenchmark}
\alias{adaBenchmark}
\title{Benchmarking AdaSampling efficacy on noisy labelled data.}
\usage{
adaBenchmark(data.mat, data.cls, data.cls.truth, s = 1, C = 50,
  sampleFactor = 1, kfolds = 5)
}
\arguments{
\item{data.mat}{a rectangular matrix or data frame that can be
coerced to a matrix, containing the
features of the dataset, without class labels. Rownames (possibly
containing unique identifiers) will be ignored.}

\item{data.cls}{a numeric vector containing class labels for the dataset
 with added noise.
Must be in the same order and of the same length as \code{data.mat}. Labels
must be 1 for positive observations, and 0 for negative observations.}

\item{data.cls.truth}{a numeric vector of true class labels for
the dataset. Must be
the same order and of the same length as \code{data.mat}. Labels must
be 1, for positive observations, and 0 for negative observations.}

\item{s}{sets the seed.}

\item{C}{sets how many times to run the classifier, for the AdaEnsemble
condition. See Description above.}

\item{sampleFactor}{provides a control on the sample size for resampling.}

\item{kfolds}{an integer for the number of folds to be used for k-fold
cross validation.}
}
\description{
\code{adaBenchmark()} allows a comparison between the performance
of an AdaSampling-enhanced SVM (support vector machine)-
classifier against the SVM-classifier on its
own. It requires a matrix of features (extracted from a labelled dataset),
and two vectors of true labels and labels with noise added as desired.
It runs an SVM classifier and returns a matrix which displays the specificity
(Sp), sensitivity (Se) and F1 score for each of four conditions:
"Original" (classifying with true labels), "Baseline" (classifying with
noisy labels), "AdaSingle" (classifying using AdaSampling) and
"AdaEnsemble" (classifying using AdaSampling in conjunction with
an ensemble of models).
}
\details{
AdaSampling is an adaptive sampling-based noise reduction method
to deal with noisy class labelled data, which acts as a wrapper for
traditional classifiers, such as support vector machines,
k-nearest neighbours, logistic regression, and linear discriminant
analysis. For more details see \code{?adaSample()}.

This function runs evaluates the AdaSampling procedure by adding noise
to a labelled dataset, and then running support vector machines on
the original and the noisy dataset. Note that this function is for
benchmarking AdaSampling performance using what is assumed to be
a well-labelled dataset. In order to run AdaSampling on a noisy dataset,
please see \code{adaSample()}.
}
\section{References}{

Yang, P., Liu, W., Yang. J. (2016) Positive unlabeled learning via wrapper-based
adaptive sampling. \emph{International Joint Conferences on Artificial Intelligence Organization.}
}

\seealso{
allows you to point to other useful resources, either on the web, \url{http://www.r-project.org}
}
