% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaSample.R
\name{adaSample}
\alias{adaSample}
\title{Implementation of AdaSampling for positive unlabelled and label noise
learning.}
\usage{
adaSample(train.mat, test.mat, cls, classifier = "knn", s = 1, C = 1,
  sampleFactor = 1)
}
\arguments{
\item{train.mat}{a rectangular matrix or data frame that can be
coerced to a matrix, containing the
features of training data, without class labels. Rownames (possibly
containing unique identifiers) will be ignored.}

\item{test.mat}{a rectangular matrix or data frame that can be
coerced to a matrix, containg the test
set, without class labels. Rownames are ignored.}

\item{cls}{a numeric vector of class labels for the training set, must be in
the same order and of the same length as \code{train.mat}. Labels must
be 1, for positive observations, and 0 for negative observations.}

\item{classifier}{the type of classifier to be used, options are
support vector machine, \code{"svm"}, k-nearest neighbour,
\code{"knn"}, logistic regression \code{"logit"}, or
linear discriminant analysis \code{"lda"}.}

\item{s}{sets the seed.}

\item{C}{sets how many times to run the classifier, C>1 induces an
ensemble learning model.}

\item{sampleFactor}{provides a control on the sample size for resampling.}
}
\description{
\code{adaSample()} applies the AdaSampling procedure to reduce noise
in the training set, and subsequently trains a classifier from
the new training set. For each row (observation) in the test set, it
returns the probabilities of it being a positive ("P) or negative
("N") instance, as a two column data frame.
}
\details{
\code{adaSample()} is an adaptive sampling-based noise reduction method
to deal with noisy class labelled data, which acts as a wrapper for
traditional classifiers, such as support vector machines,
k-nearest neighbours, logistic regression, and linear discriminant
analysis.

This process is used to build up a noise-minimized training set
that is derived by iteratively resampling the training set,
(\code{train}) based on probabilities derived after its classification.

This sampled training set is then used to train a classifier, which
is then executed on the test set. \code{adaSample()} returns a series of
predictions for each row of the test set.

Note that this function does not evaluate the quality of the model
and thus does not compare its output to true values of the test set.
To assess please see \code{adaBenchmark()}.
}
\section{References}{

Yang, P., Liu, W., Yang. J. (2016) Positive unlabeled learning via wrapper-based
adaptive sampling. \emph{International Joint Conferences on Artificial Intelligence Organization.}
}

\seealso{
allows you to point to other useful resources, either on the web, \url{http://www.r-project.org}
}
