% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca}
\alias{brca}
\title{Wisconsin Breast Cancer Database (1991)}
\format{A data frame with 683 rows and 10 variables:
\describe{
  \item{clt}{Clump thickness, 1 - 10}
  \item{ucs}{Uniformity of cell size, 1 - 10}
  \item{uch}{Uniformity of cell shape, 1 - 10}
  \item{mad}{Marginal adhesion, 1 - 10}
  \item{ecs}{Single epithelial cell size, 1 - 10}
  \item{nuc}{Bare nuclei, 1 - 10}
  \item{chr}{Bland chromatin, 1 - 10}
  \item{ncl}{Normal nucleoli, 1 - 10}
  \item{mit}{Mitoses, 1 - 10}
  \item{cla}{Class, benign or malignant}

}}
\source{
\url{https://archive.ics.uci.edu/ml/machine-learning-databases/breast-cancer-wisconsin/breast-cancer-wisconsin.data}
}
\usage{
brca
}
\description{
A cleaned version of the original Wisconsin Breast Cancer
dataset containing histological information about 683
breast cancer samples collected from patients at the
University of Wisconsin Hospitals, Madison by
Dr. William H. Wolberg between January
1989 and November 1991.
}
\references{
O. L. Mangasarian and W. H. Wolberg: "Cancer diagnosis via linear
programming", \emph{SIAM News}, Volume 23, Number 5, September 1990, pp 1 & 18.
}
\keyword{datasets}
