% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedKNN.R
\name{weightedKNN}
\alias{weightedKNN}
\title{Implementation of a feature weighted k-nearest neighbour classifier.}
\usage{
weightedKNN(train.mat, test.mat, cl, k = 3, weights)
}
\arguments{
\item{train.mat}{training data matrix, without class labels.}

\item{test.mat}{test data matrix, without class labels.}

\item{cl}{class labels for training data.}

\item{k}{number of nearest neighbour to be used.}

\item{weights}{weights to be assigned to each feautre.}
}
\description{
Implementation of a feature weighted k-nearest neighbour classifier.
}
